@@ -1,6 +1,6 @@
 {
   "name": "altinn-app-frontend",
-  "version": "3.1.1",
+  "version": "3.1.2",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -27,9 +27,9 @@
     "@material-ui/core": "~4.11.3",
     "@material-ui/icons": "~4.11.2",
     "@material-ui/pickers": "~3.3.10",
-    "@reduxjs/toolkit": "~1.5.0",
-    "ajv": "~7.2.3",
-    "ajv-formats": "~1.5.1",
+    "@reduxjs/toolkit": "~1.5.1",
+    "ajv": "~8.0.5",
+    "ajv-formats": "~2.0.2",
     "axios": "~0.21.1",
     "dot-object": "~2.1.4",
     "immutability-helper": "2.7.0",
@@ -37,10 +37,10 @@
     "lodash.isequal": "~4.5.0",
     "moment": "~2.29.1",
     "react": "~16.13.1",
-    "react-content-loader": "~6.0.2",
+    "react-content-loader": "~6.0.3",
     "react-device-detect": "~1.17.0",
     "react-dom": "~16.13.1",
-    "react-dropzone": "~11.3.1",
+    "react-dropzone": "~11.3.2",
     "react-redux": "~7.2.3",
     "react-router-dom": "~5.2.0",
     "redux": "~4.0.5",
@@ -48,16 +48,16 @@
     "redux-logger": "~3.0.6",
     "redux-saga": "~1.1.3",
     "reselect": "~4.0.0",
-    "typescript": "~4.2.3",
+    "typescript": "~4.2.4",
     "uuid": "8.3.2"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "~5.11.9",
-    "@testing-library/react": "~11.2.5",
+    "@testing-library/jest-dom": "~5.11.10",
+    "@testing-library/react": "~11.2.6",
     "@types/classnames": "~2.2.11",
     "@types/dot-object": "~2.1.2",
     "@types/enzyme": "~3.10.8",
-    "@types/jest": "~26.0.21",
+    "@types/jest": "~26.0.22",
     "@types/lodash.isequal": "~4.5.5",
     "@types/react": "~16.9.56",
     "@types/react-dom": "~16.9.12",
@@ -67,13 +67,13 @@
     "@types/redux-mock-store": "~1.0.2",
     "@types/uuid": "8.3.0",
     "altinn-designsystem": "~3.0.2",
-    "classnames": "~2.2.6",
-    "core-js": "~3.9.1",
+    "classnames": "~2.3.1",
+    "core-js": "~3.10.1",
     "cross-env": "~7.0.3",
     "css-loader": "~5.2.0",
     "enzyme": "~3.11.0",
     "enzyme-adapter-react-16": "~1.15.6",
-    "eslint": "~7.22.0",
+    "eslint": "~7.23.0",
     "eslint-config-airbnb-base": "~14.2.1",
     "fork-ts-checker-webpack-plugin": "~6.2.0",
     "fork-ts-checker-notifier-webpack-plugin": "3.0.0",
@@ -82,7 +82,7 @@
     "html-webpack-plugin": "~4.5.2",
     "jest": "~26.6.3",
     "jest-junit": "~12.0.0",
-    "mini-css-extract-plugin": "~1.3.9",
+    "mini-css-extract-plugin": "~1.4.1",
     "react-styleguidist": "~11.1.6",
     "react-test-renderer": "~16.13.1",
     "redux-mock-store": "~1.5.4",
@@ -92,7 +92,7 @@
     "svg-inline-loader": "~0.8.2",
     "terser-webpack-plugin": "~4.2.3",
     "ts-jest": "~26.5.4",
-    "ts-loader": "~8.0.18",
+    "ts-loader": "~8.1.0",
     "webpack": "~4.44.2",
     "webpack-cli": "~3.3.12",
     "webpack-dev-server": "~3.11.2"
