@@ -22,11 +22,11 @@
   "dependencies": {
     "@babel/polyfill": "~7.12.1",
     "@material-ui/core": "~4.11.3",
-    "@reduxjs/toolkit": "~1.5.0",
+    "@reduxjs/toolkit": "~1.5.1",
     "@types/domhandler": "^2.4.1",
     "axios": "~0.21.1",
-    "classnames": "~2.2.6",
-    "core-js": "~3.9.1",
+    "classnames": "~2.3.1",
+    "core-js": "~3.10.1",
     "moment": "~2.29.1",
     "react": "~16.13.1",
     "react-dom": "~16.13.1",
@@ -41,17 +41,17 @@
     "redux-logger": "~3.0.6",
     "redux-saga": "~1.1.3",
     "reselect": "~4.0.0",
-    "typescript": "~4.2.3"
+    "typescript": "~4.2.4"
   },
   "devDependencies": {
-    "@babel/core": "~7.13.10",
+    "@babel/core": "~7.13.14",
     "@babel/preset-env": "~7.13.12",
-    "@babel/preset-react": "~7.12.13",
+    "@babel/preset-react": "~7.13.13",
     "@testing-library/jest-dom": "~5.11.10",
-    "@testing-library/react": "~11.2.5",
+    "@testing-library/react": "~11.2.6",
     "@types/classnames": "~2.2.11",
     "@types/enzyme": "~3.10.8",
-    "@types/jest": "~26.0.21",
+    "@types/jest": "~26.0.22",
     "@types/react": "~16.9.56",
     "@types/react-dom": "~16.9.12",
     "@types/react-modal": "~3.12.0",
@@ -65,7 +65,7 @@
     "css-loader": "~5.2.0",
     "enzyme": "~3.11.0",
     "enzyme-adapter-react-16": "~1.15.6",
-    "eslint": "~7.22.0",
+    "eslint": "~7.23.0",
     "eslint-config-airbnb-base": "~14.2.1",
     "fork-ts-checker-notifier-webpack-plugin": "~3.0.0",
     "fork-ts-checker-webpack-plugin": "~6.2.0",
@@ -75,7 +75,7 @@
     "jest-environment-enzyme": "~7.1.2",
     "jest-enzyme": "~7.1.2",
     "jest-junit": "~12.0.0",
-    "mini-css-extract-plugin": "~1.3.9",
+    "mini-css-extract-plugin": "~1.4.1",
     "monaco-editor-webpack-plugin": "~3.0.1",
     "prettier": "~2.2.1",
     "prettier-tslint": "~0.4.2",
@@ -88,7 +88,7 @@
     "svg-inline-loader": "~0.8.2",
     "terser-webpack-plugin": "~4.2.3",
     "ts-jest": "~26.5.4",
-    "ts-loader": "~8.0.18",
+    "ts-loader": "~8.1.0",
     "webpack": "~4.44.2",
     "webpack-cli": "~3.3.12",
     "webpack-dev-server": "~3.11.2"
