@@ -22,8 +22,8 @@
   "dependencies": {
     "@material-ui/core": "~4.11.3",
     "@material-ui/icons": "~4.11.2",
-    "@reduxjs/toolkit": "~1.5.0",
-    "core-js": "~3.9.1",
+    "@reduxjs/toolkit": "~1.5.1",
+    "core-js": "~3.10.1",
     "moment": "~2.29.1",
     "react": "~16.13.1",
     "react-dom": "~16.13.1",
@@ -38,16 +38,16 @@
     "redux-logger": "~3.0.6",
     "redux-saga": "~1.1.3",
     "reselect": "~4.0.0",
-    "typescript": "~4.2.3"
+    "typescript": "~4.2.4"
   },
   "devDependencies": {
-    "@babel/core": "~7.13.10",
+    "@babel/core": "~7.13.14",
     "@babel/polyfill": "~7.12.1",
     "@babel/preset-env": "~7.13.12",
-    "@babel/preset-react": "~7.12.13",
+    "@babel/preset-react": "~7.13.13",
     "@types/classnames": "~2.2.11",
     "@types/enzyme": "~3.10.8",
-    "@types/jest": "~26.0.21",
+    "@types/jest": "~26.0.22",
     "@types/react": "~16.9.56",
     "@types/react-dom": "~16.9.12",
     "@types/react-modal": "~3.12.0",
@@ -57,12 +57,12 @@
     "@types/redux-mock-store": "~1.0.2",
     "altinn-designsystem": "2.7.12",
     "babel-jest": "~26.6.3",
-    "classnames": "~2.2.6",
+    "classnames": "~2.3.1",
     "cross-env": "~7.0.3",
     "css-loader": "~5.2.0",
     "enzyme": "~3.11.0",
     "enzyme-adapter-react-16": "~1.15.6",
-    "eslint": "~7.22.0",
+    "eslint": "~7.23.0",
     "eslint-config-airbnb-base": "~14.2.1",
     "fork-ts-checker-notifier-webpack-plugin": "~3.0.0",
     "fork-ts-checker-webpack-plugin": "~6.2.0",
@@ -72,7 +72,7 @@
     "jest-environment-enzyme": "~7.1.2",
     "jest-enzyme": "~7.1.2",
     "jest-junit": "~12.0.0",
-    "mini-css-extract-plugin": "~1.3.9",
+    "mini-css-extract-plugin": "~1.4.1",
     "prettier": "~2.2.1",
     "prettier-tslint": "~0.4.2",
     "react-test-renderer": "~16.13.1",
@@ -83,7 +83,7 @@
     "style-loader": "~2.0.0",
     "svg-inline-loader": "~0.8.2",
     "ts-jest": "~26.5.4",
-    "ts-loader": "~8.0.18",
+    "ts-loader": "~8.1.0",
     "terser-webpack-plugin": "~4.2.3",
     "webpack": "~4.44.2",
     "webpack-dev-server": "~3.11.2",
