@@ -15,9 +15,9 @@
     "start-app-development": "npm start --prefix dashboard"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "~4.19.0",
-    "@typescript-eslint/parser": "~4.19.0",
-    "eslint": "~7.22.0",
+    "@typescript-eslint/eslint-plugin": "~4.21.0",
+    "@typescript-eslint/parser": "~4.21.0",
+    "eslint": "~7.23.0",
     "eslint-config-airbnb": "~18.2.1",
     "eslint-plugin-import": "~2.22.1",
     "eslint-plugin-jsx-a11y": "~6.4.1",
