@@ -15,7 +15,7 @@
     "@material-ui/core": "~4.11.3",
     "@material-ui/icons": "~4.11.2",
     "@material-ui/lab": "4.0.0-alpha.57",
-    "@reduxjs/toolkit": "~1.5.0",
+    "@reduxjs/toolkit": "~1.5.1",
     "axios": "~0.21.1",
     "jsonpointer": "~4.1.0",
     "react": "~16.13.1",
@@ -26,14 +26,14 @@
     "@material-ui/core": "~4.11.3",
     "@material-ui/icons": "~4.11.2",
     "@material-ui/lab": "4.0.0-alpha.57",
-    "@reduxjs/toolkit": "~1.5.0",
+    "@reduxjs/toolkit": "~1.5.1",
     "@rollup/plugin-commonjs": "~17.1.0",
     "@rollup/plugin-json": "~4.1.0",
-    "@rollup/plugin-node-resolve": "~11.2.0",
+    "@rollup/plugin-node-resolve": "~11.2.1",
     "@testing-library/jest-dom": "~5.11.10",
-    "@testing-library/react": "~11.2.5",
+    "@testing-library/react": "~11.2.6",
     "@testing-library/user-event": "~12.8.3",
-    "@types/jest": "~26.0.21",
+    "@types/jest": "~26.0.22",
     "@types/react": "~16.9.56",
     "@types/react-dom": "~16.9.12",
     "@types/react-redux": "~7.1.16",
@@ -47,21 +47,21 @@
     "jest-enzyme": "~7.1.2",
     "jest-junit": "~12.0.0",
     "jsonpointer": "~4.1.0",
-    "mini-css-extract-plugin": "~1.3.9",
+    "mini-css-extract-plugin": "~1.4.1",
     "react": "~16.13.1",
     "react-dom": "~16.13.1",
     "react-redux": "~7.2.3",
     "react-scripts": "4.0.3",
     "react-styleguidist": "~11.1.6",
-    "rollup": "~2.41.5",
+    "rollup": "~2.44.0",
     "rollup-plugin-css-only": "~3.1.0",
     "rollup-plugin-peer-deps-external": "~2.2.4",
     "rollup-plugin-typescript2": "~0.30.0",
     "ts-jest": "~26.5.4",
-    "ts-loader": "~8.0.18",
-    "typescript": "~4.2.3",
-    "webpack": "~5.26.3",
-    "webpack-cli": "~4.5.0",
+    "ts-loader": "~8.1.0",
+    "typescript": "~4.2.4",
+    "webpack": "~5.31.0",
+    "webpack-cli": "~4.6.0",
     "webpack-dev-server": "~3.11.2"
   },
   "scripts": {
