@@ -24,9 +24,9 @@
     "@babel/polyfill": "~7.12.1",
     "@material-ui/core": "~4.11.3",
     "@material-ui/icons": "~4.11.2",
-    "@reduxjs/toolkit": "~1.5.0",
+    "@reduxjs/toolkit": "~1.5.1",
     "axios": "~0.21.1",
-    "core-js": "~3.9.1",
+    "core-js": "~3.10.1",
     "deepmerge": "~4.2.2",
     "react": "~16.13.1",
     "react-dnd": "~11.1.3",
@@ -42,34 +42,34 @@
     "redux-logger": "~3.0.6",
     "redux-saga": "~1.1.3",
     "reselect": "~4.0.0",
-    "typescript": "~4.2.3",
+    "typescript": "~4.2.4",
     "uuid": "8.3.2"
   },
   "devDependencies": {
-    "@babel/core": "~7.13.10",
+    "@babel/core": "~7.13.14",
     "@babel/preset-env": "~7.13.12",
-    "@babel/preset-react": "~7.12.13",
+    "@babel/preset-react": "~7.13.13",
     "@types/classnames": "~2.2.11",
     "@types/deepmerge": "~2.2.0",
     "@types/enzyme": "~3.10.8",
-    "@types/jest": "~26.0.21",
+    "@types/jest": "~26.0.22",
     "@types/react": "~16.9.56",
     "@types/react-dom": "~16.9.12",
     "@types/react-modal": "~3.12.0",
     "@types/react-redux": "~7.1.16",
     "@types/react-router-dom": "~5.1.7",
-    "@types/react-select": "~4.0.13",
+    "@types/react-select": "~4.0.14",
     "@types/react-test-renderer": "~16.9.5",
     "@types/redux-mock-store": "~1.0.2",
     "@types/uuid": "8.3.0",
     "altinn-designsystem": "2.7.12",
     "babel-jest": "~26.6.3",
-    "classnames": "~2.2.6",
+    "classnames": "~2.3.1",
     "cross-env": "~7.0.3",
     "css-loader": "~5.2.0",
     "enzyme": "~3.11.0",
     "enzyme-adapter-react-16": "~1.15.6",
-    "eslint": "~7.22.0",
+    "eslint": "~7.23.0",
     "eslint-config-airbnb-base": "~14.2.1",
     "html-loader": "~1.3.2",
     "html-webpack-plugin": "~4.5.2",
@@ -77,7 +77,7 @@
     "jest-environment-enzyme": "~7.1.2",
     "jest-enzyme": "~7.1.2",
     "jest-junit": "~12.0.0",
-    "mini-css-extract-plugin": "~1.3.9",
+    "mini-css-extract-plugin": "~1.4.1",
     "prettier": "~2.2.1",
     "prettier-tslint": "~0.4.2",
     "react-test-renderer": "~16.13.1",
@@ -88,7 +88,7 @@
     "style-loader": "~2.0.0",
     "svg-inline-loader": "~0.8.2",
     "ts-jest": "~26.5.4",
-    "ts-loader": "~8.0.18",
+    "ts-loader": "~8.1.0",
     "terser-webpack-plugin": "~4.2.3",
     "webpack": "~4.44.2",
     "webpack-dev-server": "~3.11.2",
