@@ -533,7 +533,7 @@ export function validateFormDataForLayout(
           [errorParams],
         );
 
-        const dataBindingName = processDataPath(error.dataPath);
+        const dataBindingName = processInstancePath(error.instancePath);
         mapToComponentValidations(layoutKey, layout, dataBindingName, errorMessage, result.validations);
       }
     });
@@ -542,7 +542,7 @@ export function validateFormDataForLayout(
   return result;
 }
 
-export function processDataPath(path: string): string {
+export function processInstancePath(path: string): string {
   let result = path.startsWith('.') ? path.slice(1) : path;
   result = result.replace(/']\['/g, '.').replace(/\['/g, '').replace(/']/g, '');
   return result;
